run mdformat and bump ruff pre-commit (#1036)
We use mdformat to automatically format md files in bionemo-recipes, as
part of the prep work to merging these repos we could add mdformatting
to the existing bionemo-framework repo.

This adds 
```
  - repo: https://github.com/executablebooks/mdformat
    rev: 0.7.17             # Use the latest stable version
    hooks:
      - id: mdformat
        additional_dependencies:
        - mdformat-tables
        - mdformat-gfm
        - mdformat-black
        - mdformat-frontmatter
```

to .pre-commit-config.yaml, which automatically formats markdown files
(and code blocks within markdown files). It also adds an
`.mdformat.yaml` file to customize the mdformat behavior.

+++ sub-packages/bionemo-geneformer/src/bionemo/geneformer/scripts/train_geneformer.py
@@ -343,10 +343,10 @@ def main(
                 max_steps=num_steps,
                 # minimum learning rate is 1/100th of the initial learning rate, so eg lr=1e-3 -> min_lr=1e-5
                 min_lr=lr / 100,
-                warmup_steps=int(math.ceil(num_steps * cosine_rampup_frac)),
+                warmup_steps=math.ceil(num_steps * cosine_rampup_frac),
                 interval="step",
                 monitor="val_loss",
-                constant_steps=int(math.ceil(num_steps * cosine_hold_frac)),
+                constant_steps=math.ceil(num_steps * cosine_hold_frac),
             ),
         ),
     )
